be/requirement.txt
langgraph
langchain
langchain-openai
langchain-google-genai
pydantic
fastapi
uvicorn
python-dotenv
